Add files via upload
click-outside has a function that when clicks on any areas where outside of images, those images disappear. To build this website in js, it required to create 2 more extra div tags that covers basic layer of website which are modal-outer and modal-inner. Also it required to access to information such as image src and description via javascript so then that the information can be applied into buttons, so when buttons are clicked, image with its description pops up.